requirements.txt
streamlit
streamlit_authenticator
openai
requests
pyyaml
streamilt_extras
